Refactoring on all the major components increased minimum sdk version to 21 introduces dagger 2.11 dependency injection Profile fragments and settings implemented with data binding library and MVVM added livedata, live preferences and lifecycle aware viewmodels code cleanup

+++ app/src/main/java/com/telenav/osv/listener/ShutterListener.java
@@ -6,5 +6,5 @@ package com.telenav.osv.listener;
  */
 public interface ShutterListener {
 
-  void takeSnapshot(float distance);
+  void takeSnapshot(double distance);
 }
